Make the aggregate version consistent for handling multiple commands without reconsistion from the message repository.
